Recognize non-ASCII escape sequences
As suggested in section 21.7.15 of the elisp manual, handle keyboard
events internally as vectors across the board, rather than preferring
strings when possible. This commit, in combination with the one before
it, fixes syl20bnr/evil-escape#51: `this-command-keys' returns a vector
for most key sequences, so (for instance) the expression

  (equal (this-command-keys) "π")

is always nil.